3️⃣ 로그인 리팩터링 (#55)
* misc: 계획을 적었다. (everybody has a plan until they get punched in the face.)

* feat: extract AuthenticationConverter interface from interceptors.

* test(session-auth-converter): added test for SessionAuthenticationConverter to convert request to auth-token

* feat(session-auth-converter): Implemented SessionAuthenticationConverter

* refactor(session-auth-interceptor): make Interceptor uses authenticationConverter for converting request to token.

* test(token-auth-conveter): added test for TokenAuthenticationConverter to convert basic auth to token.

* feat(token-auth-conveter): implemented TokenAuthenticationConverter

* refactor: fixed test. fixed interceptor

* refactor: extract superclass from interceptors

* feat: extract UserDetail interface from LoginMember

* feat: extract UserDetailsService interface from CustomerUserDetailsService

* refactor: converted existing classes to subclass of interface in order to remove circular dependencies.

* refactor: removed all member dependencies from auth package.

* feat: added AuthenticatedUserDetailsMethodArgumentResolver to resolve AuthenticationPrincipal annotation

* bugfix: fixed argument resolver error...

* feat: added feature(modify/update) for Member

* refactor: removed unused modules and tests after migration complete.

+++ src/main/java/nextstep/subway/member/ui/MemberController.java
@@ -1,10 +1,11 @@
 package nextstep.subway.member.ui;
 
-import nextstep.subway.auth.infrastructure.SecurityContextHolder;
+import nextstep.subway.auth.application.UserDetail;
+import nextstep.subway.auth.domain.AuthenticationPrincipal;
 import nextstep.subway.member.application.MemberService;
-import nextstep.subway.member.domain.LoginMember;
 import nextstep.subway.member.dto.MemberRequest;
 import nextstep.subway.member.dto.MemberResponse;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -31,20 +32,28 @@ public class MemberController {
     }
 
     @GetMapping("/members/me")
-    public ResponseEntity<MemberResponse> findMemberOfMine() {
-        final LoginMember loginMember = (LoginMember) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        return ResponseEntity.ok(new MemberResponse(loginMember.getId(), loginMember.getEmail(), loginMember.getAge()));
+    public ResponseEntity<MemberResponse> findMemberOfMine(@AuthenticationPrincipal UserDetail userDetail) {
+        return ResponseEntity.ok(new MemberResponse(userDetail.getId(), userDetail.getEmail(), userDetail.getAge()));
     }
 
     @PutMapping("/members/{id}")
-    public ResponseEntity<MemberResponse> updateMember(@PathVariable Long id, @RequestBody MemberRequest param) {
-        memberService.updateMember(id, param);
-        return ResponseEntity.ok().build();
+    public ResponseEntity<MemberResponse> updateMember(
+            @PathVariable Long id,
+            @RequestBody MemberRequest param,
+            @AuthenticationPrincipal UserDetail userDetail) {
+        if (id.equals(userDetail.getId())) {
+            memberService.updateMember(id, param);
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
     @DeleteMapping("/members/{id}")
-    public ResponseEntity<MemberResponse> deleteMember(@PathVariable Long id) {
-        memberService.deleteMember(id);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<MemberResponse> deleteMember(@PathVariable Long id, @AuthenticationPrincipal UserDetail userDetail) {
+        if (id.equals(userDetail.getId())) {
+            memberService.deleteMember(id);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 }
